delete fadeInUp animations in separate components
