--- src/components/AdvantagesList/variants.ts
@@ -1,15 +1,0 @@
-export const fadeInUp = {
-  hide: {
-    opacity: 0,
-    y: 50,
-  },
-  show: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      staggerChildren: 0.3,
-      type: 'spring',
-      stiffness: 35,
-    },
-  },
-};
