+++ src/components/Statistics/Statistics.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import stats from '@/data/statistics.json';
 import { StatisticsProps } from './types';
-import { fadeInUp } from './variants';
+import { fadeInUp } from '@/variantsCommon/fadeInUp';
 
 export const Statistics: React.FC<StatisticsProps> = ({ className }) => {
   return (
