+++ src/components/TimelineList/TimelineList.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import css from './Timeline.module.css';
 import data from '@/data/timeline.json';
 import { TimelineListItem } from '../TimelineListItem';
-import { fadeInUp } from './variants';
+import { fadeInUp } from '@/variantsCommon/fadeInUp';
 
 export const TimelineList: React.FC = () => {
   return (
